requirements.txt
streamlit
yt-dlp
requests
pillow
